requirements.txt
python-dotenv
transformers
datasets
torch
sentence-transformers
openpyxl
pandas
langchain-community
faiss-cpu
langchain-text-splitters
qdrant-client
langchain
langchain-community
langchain-core
langchain-openai
langgraph